Ultimate Tic-Tac-Toe Game 2-Player
This is my Ultimate tic-tac-toe game that uses 2-player.
This game is a large tic-tac-toe board, with smaller tic-tac-toe boards in each of the squares. Turns are taken in
the small boards. When a player makes a turn in the small tic-tac-toe board, then the next player plays in the small board that corresponds to the square the previous player played. For Example: Player "X" plays their turn in the "middle-right" board, and in that board they play the square "top-middle". Now Player "O" is forced to play in the "top-middle" board and can click on any of the squares in that board. If,"X" plays a square, where the corresponding board is already won, then "O" can play their turn on any square, from any un-won boards. Those are the rules, the goal is to create a three-in-a-row pattern in the large board and win the game.
Good Luck